Move Blog Carousel to Components
